Change name of FAN1 from heatbreak_cooling_fan to controller_fan
On the Ender 3 Pro, FAN1 is connected to the controller/mainboard fan, not to the hotend fan.  The hotend fan is hardwired to 24v when the printer is on.  One could argue that a controllable hotend fan is more useful, but that's not how they're configured from the factory.

I suspect other Ender 3s are the same way, but I don't have others to actually test so I didn't want to propose changes for devices I don't have.  I'd be glad to change the others if you want though.